refactor: Coordinate 패키지 묶기
