+++ src/main/java/chessgame/domain/coordinate/Inclination.java
@@ -1,4 +1,4 @@
-package chessgame.domain.piecetype;
+package chessgame.domain.coordinate;
 
 import java.util.Arrays;
