+++ src/main/java/chessgame/domain/piece/EmptyPiece.java
@@ -1,6 +1,6 @@
 package chessgame.domain.piece;
 
-import chessgame.domain.piecetype.Coordinate;
+import chessgame.domain.coordinate.Coordinate;
 import chessgame.domain.piecetype.PieceType;
 
 import java.util.Optional;
