+++ src/main/java/chessgame/domain/piecetype/Bishop.java
@@ -1,9 +1,12 @@
 package chessgame.domain.piecetype;
 
+import chessgame.domain.coordinate.Coordinate;
+import chessgame.domain.coordinate.Inclination;
+
 import java.util.List;
 
-import static chessgame.domain.piecetype.Inclination.MINUS_ONE;
-import static chessgame.domain.piecetype.Inclination.ONE;
+import static chessgame.domain.coordinate.Inclination.MINUS_ONE;
+import static chessgame.domain.coordinate.Inclination.ONE;
 
 public class Bishop implements PieceType {
