+++ src/main/java/chessgame/domain/piecetype/PieceType.java
@@ -1,5 +1,7 @@
 package chessgame.domain.piecetype;
 
+import chessgame.domain.coordinate.Coordinate;
+
 public interface PieceType {
 
     boolean isReachableByRule(Coordinate startCoordinate, Coordinate endCoordinate);
