+++ src/test/java/chessgame/domain/board/BoardInitialImageTest.java
@@ -1,7 +1,7 @@
 package chessgame.domain.board;
 
+import chessgame.domain.coordinate.Coordinate;
 import chessgame.domain.piece.Piece;
-import chessgame.domain.piecetype.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
