+++ src/test/java/chessgame/domain/piecetype/PawnTest.java
@@ -1,5 +1,6 @@
 package chessgame.domain.piecetype;
 
+import chessgame.domain.coordinate.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
