+++ src/test/java/chessgame/domain/piecetype/QueenTest.java
@@ -1,5 +1,6 @@
 package chessgame.domain.piecetype;
 
+import chessgame.domain.coordinate.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
